extract admin sign-in logic into helper functions
